change useTheme to another file
